Mark service-compos as system_ext_specific
Although service-compos is not installed in /system_ext, it is an apex
system server jar. These jars are preopted and the files are installed
outside the apex. All such files were previously installed in /system/,
but with https://r.android.com/3288083, these will be installed in the
same partition as the parent apex.

Test: Verified that
apex@com.android.compos@javalib@service-compos.jar@classes.odex is
installed in /system_ext and not in /system (or /system/system_ext if
system_ext does not exist)

Bug: 369678122
Change-Id: Id0e7b462881bc24c94e60102d9c67b48a0787243