Use statically assigned IP for the VM
Ideally, we will use mDNS to resolve the IP address of the Debian OS VM.
However, currently, mDNS doesn't seem to work on tethered network.

Until then, let's statically assign IP to the VM. The assumption here is
that no multiple VMs with network access exist, which is a valid
assumption at least for now.

The Terminal app now doesn't wait to the message from the guest agent.
Instead, it waits for the static IP to become reachable, and also
retries when the terminal service fails to load.

In addition, the WebView is invisible in the beginning and becomes
visible only after the page is fully loaded.

Bug: N/A
Test: run the app.
Change-Id: Ibf084a9c6e15835394bf404ed1ddbcbd1e3b463c