Set hint for some UI elements
In the ttyd UI, the edit box (cursor) is the only elements that supports
action. However, it's too tiny, so users may find it difficult to set
the a11y focus on it using explore-by-touch. Even with gestures (swipe
next/prev), the cursor is hard to reach because it's located in the very
end of the traversal order; user has to go through 30-40 elements each
of which is for a line in the terminal.

To fix this issue, ttyd seems to have implemented a feature that double
tapping anywhere in the terminal UI screen activates the edit box
directly.

This however may confuse accessibility users because, they didn't hear
that double tapping action is provided. This is b/376827536.

We tried hard to fix this issue correctly; by adding ACTION_CLICK to the
other UI elements, and many more attempts, but all failed. For example,
adding ACTION_CLICK made talkback to announce "double-tap to activate",
but actually double-tapping the screen didn't lead user to the cursor.
This is probably because the click action is consumed by the other UI
elements, not the edit box.

To work around these limitations, this change uses hint text to let the
users know that double tapping will go to the cursor directly.

Bug: 376827536
Test: follow the instruction.
Change-Id: I57a3f3673209c10fdd5441229da6c0e2bf1a89dd