build/debian: Look for kernel source in Debian security as well
The source packages for kernel version 6.1.119-1 used for building
our custom kernel were moved to the Debian security channel,
which resulted in the build.sh script breaking with a 404 error.

In this change, we first attempt to grab the kernel source from the
security channel and then fall back to the regular channel.

Bug: 390477424
Test: ./build_in_container.sh -k
Test: ./build_in_container.sh -a aarch64 -k

Change-Id: I555fbbd1c2a9455021c87654a433a1f676468e99