pvmfw: rollback: Remove special RBP for Trusty VM
Use deferred RBP for Trusty VM, which is identical to the RBP solution
introduced by aosp/3367779 with the extra constraint that the RB index
should be non-null (not bad to ensure).

This prepares for fixing the currently-supported Capabilities, which
have actually been defined per use-case, into an interface that can be
stabilized for Treble in Android 16 as an upcoming change will introduce
a separate AVB property to select the RBP solution and both users of
SecretkeeperProtection and TrustySecurityVm are expected to use a
unified one.

Bug: 391620545
Test: m pvmfw_bin && TH
Change-Id: I852107dee56ce257a03fbec3e7ae706c1f7c93a8