pvmfw: rollback: Force fixed RBP for RemoteAttest
If an image pretends to be the RKP VM ("Capability::RemoteAttest"),
force that it goes through the intended fixed-index rollback solution
instead of potentially enabling the use of other RB solutions (such as
deferred) and still adding the RKP VM marker to the DICE chain. This
effectively makes pvmfw ignore the other use-case-specific Capability
values when it recognizes the RKP VM.

More fundamentally, this introduces the logic that pvmfw should always
(and only) check "special" VMs with the appropriate RBP method, an
approach that should survive converting Capability::RemoteAttest into a
separate (re-usable) identifying property (b/378673494) but also adding
support for new fixed RBP methods such as "fixed digest" (b/392628867).

Bug: 377276983
Bug: 391620545
Bug: 392628867
Test: m pvmfw_bin
Change-Id: Ia25a6265ed3f0466ff8bd22411636e5f7a9e5d60