pvmfw: fdt: Reject partial initrd & kernel ranges
As those ranges are described by pairs of properties (as opposed to the
now-standard <reg> property), an input DT could contain only one, making
it impossible to fully describe the range.

Previously, this would result in the range being considered as missing
from the DT. For the kernel range, pvmfw would ultimately abort the boot
as it requires a payload but for the initrd range, it would consider
that the initrd was missing, silently stripping the (single) property
from the DT and potentially booting the guest successfully.

Instead, treat these corner cases are errors during DT validation.

Test: m pvmfw
Change-Id: Ic70945e84e8bb83b42acfe2e55895267e15d40f6