virtmgr: don't check label for vendor initiated VMs
Checking for a prefix on the label is a bad practice and doesn't enforce
the relevant constraint (that the file isn't writable), so just skip the
check altogether for now.

Note that we have a separate check that verifes the files must come from
the vendor partition and that isn't changing in this commit.

Bug: 397499254
Bug: 393087663
Test: m
Change-Id: Iaa3201a78effb63184e80c51bf4c2550298ea299