Refine CrosvmConfig balloon filed setting logic
This CL makes CrosvmConfig's balloon field represent the correct balloon
enable status. Previously, whether to use balloon or not is defined in
run_vm() function. The vm may disable balloon even if config.balloon is
true. This logic is moved to before initializing CrosvmConfig.

Addtionally, this change introduces an API isMemoryBalloonEnabled() to
check if memory ballooning is enabled for a given VM.

Bug: b/392791968
Test: test with https://r.android.com/3502917
Change-Id: I79f722800aa9a6ed90e8455d185398bc4292b094